pyproject.toml
[project]
name = "dirac-bench"
version = "0.1.0"
description = "Benchmark suite for QCI Dirac-3 quantum solver on Motzkin-Straus max-clique problems"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "scipy",
    "networkx",
    "matplotlib",
    "python-dotenv",
    "eqc-models",
    "qci-client>=4.5",
    "eqc-direct>=1.0.7",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dirac_bench"]
